+++ esp-hal-embassy/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Removed the TIMG and SYSTIMER time-drivers, replaced by a generic time-driver taking `OneShotTimer<ErasedTimer>` (#1753)
 
 ### Fixed