+++ esp-wifi/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `esp_wifi::initialize` no longer requires running maximum CPU clock, instead check it runs above 80MHz. (#1688)
 - Rename `set_mode` to `set_protocol`, also available in esp-now API (#1742)
+- `esp_wifi::initialize` now takes a `PeriodicTimer<ErasedTimer>` (#1753)
 
 ### Removed